% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcsin_trans.R
\name{arcsin_trans}
\alias{arcsin_trans}
\title{Arcsin data transformation for single-cell cytometry data}
\usage{
arcsin_trans(
  input_folder,
  cofactor = 5,
  output_folder = ".",
  channel.plot = FALSE
)
}
\arguments{
\item{input_folder}{Path to folder containing fcs files.}

\item{cofactor}{Numeric. The value to be transformed is divided by this cofactor before arcsin transformation.
This value controls the extent of the linear region around 0. Default is 5.}

\item{output_folder}{Path to folder to save the transformed fcs files.}

\item{channel.plot}{Logical (default is FALSE). Plot each channel before and after transformation for a randam fcs file in the dataset for visual inspection of transformation. Plots are exported to \code{output_folder}.}
}
\value{
Arcsin transformed fcs files. And, histograms of all channels of one randomly chosen file.
}
\description{
Performs data transformation to cytometry single-cell data.
}
