% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_drop.R
\name{cluster_drop}
\alias{cluster_drop}
\title{Drop clusters of single cell cytometry data}
\usage{
cluster_drop(
  SE,
  clusters,
  reassign = FALSE,
  output_folder = NULL,
  SE_name = NULL
)
}
\arguments{
\item{SE}{SummarizedExperiment object or path to SummarizedExperiment object.}

\item{clusters}{Vector of the cluster IDs that needs to be removed.}

\item{reassign}{Logical. If TRUE, the cluster IDs are reassigned new IDs starting from 1.}

\item{output_folder}{Path of output directory to receive the subsetted SummarizedExperiment.}

\item{SE_name}{Name of the new subsetted SummarizedExperiment.}
}
\description{
Remove specific clusters from SummarizedExperiment object
}
