% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_clust_colsPie.R
\name{cyto_clust_colsPie}
\alias{cyto_clust_colsPie}
\title{Stacked columns and pie chart for single cell cytometry data}
\usage{
cyto_clust_colsPie(SE, groups = "groups", output_folder = ".")
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object or path to \code{SummarizedExperiment} object.}

\item{groups}{Experimental groups to generate the plots for.}

\item{output_folder}{Path to folder to receive pdf file of the plots.}
}
\value{
Saved pdf file containing the plots.
}
\description{
Generates stacked column and pie charts plots for cell-cluster abundance in clustered single-cell cytometry data.
}
