% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_clust_corr.R
\name{cyto_clust_corr}
\alias{cyto_clust_corr}
\title{Calculate and plot correlation of cluster abundance.}
\usage{
cyto_clust_corr(
  SE,
  assay_name = "exprs",
  groups = NULL,
  output_folder = ".",
  ...
)
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object or path to \code{SummarizedExperiment} object.}

\item{assay_name}{A character vector of length 1 indicates the assay to be used in plotting the UMAP. By default, it is \code{"exprs"}.}

\item{groups}{Experimental groups to generate the plots for.}

\item{output_folder}{Path to folder to receive pdf file of the plots.}

\item{...}{Extra arguments passed to \code{\link[corrplot]{corrplot}}}
}
\value{
correlation plots between cell-clusters for clustered cytometry single cell data.
}
\description{
Calculates and plots correlation of cluster abundance in clustered single-cell cytometry data.
}
