% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_clust_limma.R
\name{cyto_clust_limma}
\alias{cyto_clust_limma}
\title{Differential expression of biomarkers for clustered single-cell cytometry data using limma}
\usage{
cyto_clust_limma(
  SE,
  contrast,
  groups = "groups",
  exp_subject = NULL,
  paired = FALSE,
  output_folder = "."
)
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object or path to \code{SummarizedExperiment} object.}

\item{contrast}{A character vector of length 2, specify two levels of "groups" variable to build contrast matrix. \code{(contrast[1]-constrast[2])}.}

\item{groups}{Experimental groups to generate the plots for. It has to be a variable in the data i.e. in \code{rowdata()}. Levels of this variable will be compared.}

\item{exp_subject}{Default is NULL. If the experimental design is paired, this variable will be used in the formula to build paired design matrix.}

\item{paired}{Logical (default is FALSE). If TRUE a paried experimental design is performed. The pairing is done on the exp_subject variable.}

\item{output_folder}{Path to folder to receive limma output table.}
}
\value{
A table (csv file) with log fold change and statistical values for the comparison between the two contrasting groups using limma.
}
\description{
Performs differential expression analysis to clustered cytometry data using limma (markers labelled as state in colData()).
}
\seealso{
\href{https://bioconductor.org/packages/release/bioc/html/limma.html}{limma}.\href{https://github.com/lmweber/diffcyt}{diffcyt}
}
