% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_clust_violin.R
\name{cyto_clust_violin}
\alias{cyto_clust_violin}
\title{Violin plots for single-cell cytometry data}
\usage{
cyto_clust_violin(
  SE,
  assay_name = "exprs",
  marker_list = "all",
  groups = NULL,
  output_folder = "."
)
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object or path to \code{SummarizedExperiment} object.}

\item{assay_name}{Name of \code{assay()} slot to be used in generating the violin plots.}

\item{marker_list}{Character vector containing the name of the markers to generate violin plots for. It must be included
in \code{clean_name} column of \code{colData(SE)}}

\item{groups}{Experimental groups to generate the plots for.}

\item{output_folder}{Path to folder to receive pdf file of the violin plots.}
}
\value{
Saved pdf file containing violin plots.
}
\description{
Generates violin plots of marker expression intensity on cellular clustered data.
}
