% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_diff_exprs.R
\name{cyto_diff_exprs}
\alias{cyto_diff_exprs}
\title{Differential expression analysis of biomarkers in clustered single-cell cytometry data}
\usage{
cyto_diff_exprs(
  SE,
  groups = "groups",
  assay_name = "exprs",
  exp_subject = "sample_id",
  count_threshold = 100,
  output_folder = ".",
  num_boot = 1000,
  paired = FALSE
)
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object or path to \code{SummarizedExperiment} object.}

\item{groups}{Experimental groups to generate the plots for. It has to be a variable in the data i.e. in \code{rowdata()}. Levels of this variable will be compared.}

\item{assay_name}{A character vector of length 1 indicates the assay to be used in plotting the UMAP. By default, it is \code{"exprs"}.}

\item{exp_subject}{A character vector of length 1 indicates the experimental subject e.g patient_id, sample_id..etc.}

\item{count_threshold}{An Integer (default is 100). Clusters with cell count that are equal or less than this value will be excluded from the analysis.}

\item{output_folder}{Path to folder to receive pdf file of the plots and tables.}

\item{num_boot}{Numeric. Indicates the number of bootstrapping runs for the bootstrapped generalized linear model. Only used for the unpaired experimental design.}

\item{paired}{logical (default is FALSE). Indicate whether the experiment design has a pairing variable. If TRUE, the exp_subject will be used as the pairing variable.}
}
\value{
Forest plot for marker of interest on each cluster available in the data.
Table of coefficients and significance values.
}
\description{
Performs GLM modeling to clustered single-cell cytometry data. Treats selected biomarkers as explanatory variables and the experimental conditions as response variable.
}
\seealso{
\href{https://github.com/ChristofSeiler/CytoGLMM}{CytoGLMM}.
}
