% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_sample_barplot.R
\name{cyto_samples_barplot}
\alias{cyto_samples_barplot}
\title{Single cell cytometry data visualization}
\usage{
cyto_samples_barplot(SE, samples = "sample_id", output_folder = ".")
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object or path to \code{SummarizedExperiment} object.}

\item{samples}{Character. It must be in \code{colnames} \code{(rowData())}, indicates the sample name/ID.}

\item{output_folder}{Path to folder to receive pdf file of the columns plot.}
}
\value{
stacked columns plot for cluster proportions within all samples in the dataset.

Barplot for cluster proportions in each sample in the data, and csv file with this data.
}
\description{
Plot a stacked columns plot for the proportion of clusters in each sample in clustered cytometry data.
}
