% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_umap.R
\name{cyto_umap}
\alias{cyto_umap}
\title{Visualize phenograph clusters of cytometry single-cells data on UMAP axes.}
\usage{
cyto_umap(
  SE,
  output_folder,
  groups = NULL,
  marker_int = TRUE,
  marker_list = "all",
  assay_name = "exprs",
  plotwrap = FALSE,
  DS = 5000,
  label_size = 5,
  dot_size = 0.1,
  width = 8,
  height = 8,
  ...
)
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object or path to  \code{SummarizedExperiment} object.}

\item{output_folder}{Path to folder to receive the generated plots.}

\item{groups}{Experimental groups to generate plots for.}

\item{marker_int}{Logical. If TRUE, UMAPs for markers intensity are generated.}

\item{marker_list}{Character vector containing the name of the proteins of interest to generate a UMAP for. It must be included
in \code{clean_name} column of \code{colData(SE)}}

\item{assay_name}{A single character vector indicates the assay to be used in plotting the UMAP. By default, it is \code{"exprs"}.}

\item{plotwrap}{Logical. If TRUE, a universal plot of the UMAPs of the proteins listed in \code{marker_list} is generated and saved.}

\item{DS}{Numeric. Number of cells to downsample for visualization only. If \code{groups} is NULL, this value will be used to downsample the whole dataset.
If groups is not NULL, this value will be used to downsample from each group provided.}

\item{label_size}{Numeric. Controls the cluster label size.}

\item{dot_size}{Numeric. Controls the plot dot size.}

\item{width}{The plot width in inches.}

\item{height}{The plot height in inches.}

\item{...}{Extra arguments to \code{ggplot()}}
}
\description{
Plot phenograph clusters of cytometry single-cells data on UMAP axes.
}
