% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_workflow.R
\name{cyto_workflow}
\alias{cyto_workflow}
\title{Single-cell cytometry data workflow}
\usage{
cyto_workflow(
  input_folder,
  SE = NULL,
  output_folder,
  marker_info,
  experiment_info,
  batch_correct = FALSE,
  markers_correct = NULL,
  xdim = 8,
  ydim = 8,
  k = 30,
  n = 15,
  umap_method = NULL,
  conda_env = NULL,
  assay_name = "exprs",
  plot_clusters = TRUE,
  DS = 5000,
  groups = NULL,
  filter = NULL,
  exp_subject = "sample_id",
  plot_samples_prop = TRUE,
  samples = "donor_id",
  plot_clusters_violin = TRUE,
  plot_cluster_corr = TRUE,
  plot_cluster_colsPie = TRUE,
  plot_cluster_hm = TRUE,
  diff_exprs = TRUE,
  count_threshold = 100,
  num_boot = 1000,
  diff_abund = TRUE,
  diff_limma = TRUE,
  contrast = NULL,
  paired = FALSE,
  seed = 7439815,
  ...
)
}
\arguments{
\item{input_folder}{Path to folder containing fcs files.}

\item{SE}{SummarizedExperiment object or path to SummarizedExperiment object.}

\item{output_folder}{Path to folder to receive output.}

\item{marker_info}{Path to marker_info table; one of the csv files generated by \code{\link{metadata_tables}}.}

\item{experiment_info}{Path to experiment_info table; one of the csv files generated by \code{\link{metadata_tables}}. It must have a "sample_id". And a
column "batch", if batch_correct is TRUE.}

\item{batch_correct}{Logical. If TRUE, batch correction using 'cyCombine' package.}

\item{markers_correct}{Character vector specifying the markers to be corrected for batch effect. It is recommended to run \code{\link[cyCombine]{detect_batch_effect}} to
define and evaluate markers that needs batch correction.}

\item{xdim}{The x-dimension size of the SOM. See \code{\link[cyCombine]{batch_correct}}.}

\item{ydim}{The y-dimension size of the SOM. See \code{\link[cyCombine]{batch_correct}}.}

\item{k}{Integer. k paramter for phenograph clustering. See \code{\link{k_sweep}}}

\item{n}{Integer. The size of local neighborhood (n_neigbors paramter of UMAP). See \code{\link[umap]{umap}}}

\item{umap_method}{If set to \code{umap-learn}, python umap-learn module will be used via reticulate. If \code{NULL}, R implementation of umap is used. \code{umap-learn} is recommended for large data sets.}

\item{conda_env}{Path to conda environment that contain umap module. IT can be create through conda install. Please check \href{https://anaconda.org/conda-forge/umap-learn}.}

\item{assay_name}{A single character vector indicates the assay to be used in plotting the UMAP. By default, it is \code{exprs}.}

\item{plot_clusters}{Logical. If TRUE, \code{\link{cyto_umap}} is called to plot phenograph clusters on umap axes.}

\item{DS}{Integer. Specifies the number of cell to downsample to for the umap plots. If groups are provided, this number will be used to downsample from each group.}

\item{groups}{Character vector specifying the experimental group e.g c("control, "treated").}

\item{filter}{A character vector of all the levels in 'groups' that need to be filtered out of the analysis. Default is NULL.}

\item{exp_subject}{Character vector specifying the experimental subject e.g c("mouse1", "mouse2").}

\item{plot_samples_prop}{Logical. If TRUE, \code{\link{cyto_samples_barplot}} is called to plot cluster proportion in each sample in "sample_id".}

\item{samples}{Character. Represents the column name in the metadata to plot the sample barplot for.}

\item{plot_clusters_violin}{Logical. If TRUE, \code{\link{cyto_clust_violin}} is called to plot marker intensity per cluster.}

\item{plot_cluster_corr}{Logical. If TRUE, \code{\link{cyto_clust_corr}} is called to plot correlation matrix between clusters in samples and groups.}

\item{plot_cluster_colsPie}{Logical. If TRUE, \code{\link{cyto_clust_colsPie}} is called to plot proportions of clusters in each group in a column plot and pie chart.}

\item{plot_cluster_hm}{Logical. If TRUE, \code{\link{cyto_clust_hm}} is called to plot heatmap for the median marker intensity per cluster.}

\item{diff_exprs}{Logical. If TRUE, \code{\link{cyto_diff_exprs}} is called to perform differential expression test using 'CytoGLMM'.}

\item{count_threshold}{Integer. Clusters with cell count that are equal or less than this value will be excluded from the CytoGLMM differential expression analysis.}

\item{diff_abund}{Logical. If TRUE, \code{\link{cyto_clust_abundance}} is called to perform differential abundance test using 'diffcyt' and 'edgeR'.}

\item{diff_limma}{Logical. If TRUE, \code{\link{cyto_clust_limma}} is called to perform differential expression test using 'diffcyt' and' limma'.}

\item{contrast}{A character vector of length 2, specify two levels of "groups" variable to build contrast matrix. \code{(contrast[1]-constrast[2])}.}

\item{paired}{Logical. If TRUE, paired experimental design is taken in consideration for all the statistical tests. The pairing is done on the column provided in \code{exp_subject}.}

\item{seed}{Setting random seed for results reproducibility. It can be NULL.}

\item{...}{Additional arguments to any of the functions used in the wrapper.}
}
\value{
A variety of R objects, tables and figures for cytometry data analysis exported to 'output_folder'.
}
\description{
This is a wrapper function around package 'Cytonalyze' to automate cytometry data analysis.
}
