% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDDsample.R
\name{DDDsample}
\alias{DDDsample}
\title{Density-dependent Downsampling}
\usage{
DDDsample(
  input_folder,
  output_folder,
  marker_info,
  experiment_info,
  arcsinh_cofactor = NULL,
  density_to_exclude = 0.01,
  density_to_preserve = 0.03,
  percent_events_keep = NULL,
  number_events_keep = NULL,
  ...
)
}
\arguments{
\item{input_folder}{Path to folder containing raw fcs files.}

\item{output_folder}{Path to folder to receive downsample fcs files.}

\item{marker_info}{Path to marker_info; one of the csv files generated by \code{\link{metadata_tables}}. Only markers that are labelled "type"
will be used to calculate cell density.}

\item{experiment_info}{Path to experiment_info; one of the csv files generated by \code{\link{metadata_tables}}.}

\item{arcsinh_cofactor}{Cofactor value for the data arcsinh transformation. Default value is 5, which is the typically recommended data for mass cytometry.
Recommended value for flow cytometry is 150.}

\item{density_to_exclude}{Cells below this density value will be excluded. Default is 0.01 percentile.}

\item{density_to_preserve}{Cells between this density value and the \option{density_to_exclude} value will be kept untouched.
Cells above this value will be downsampled till their density matches this value. Default is 0.3 percentile.}

\item{percent_events_keep}{Percent of cells per file to keep after the downsampling. e.g the lowest 50\% in density.
A numeric from 0 to 1.}

\item{number_events_keep}{Number of events per file to keep. This will downsample approximately equal number of events per file.}

\item{...}{Extra arguments passed to \code{\link[spade]{SPADE.addDensityToFCS}}.}
}
\value{
Two folders; the first contains all the input fcs files with additional parameter added "density" added to the list of parameters.
The second folder contains the downsampled fcs files.
}
\description{
Downsample individual fcs files based on calculating local density for each single cell.
For more details, see \code{\link[spade]{SPADE.addDensityToFCS}}.
}
\note{
For more information; see \href{https://www.rdocumentation.org/packages/spade/versions/1.20.0}{spade}
}
\references{
\href{https://www.nature.com/articles/nbt.1991}{Extracting a cellular hierarchy from high-dimensional cytometry data with SPADE}
}
\seealso{
\code{\link[spade]{SPADE.addDensityToFCS}}, \code{\link[spade]{SPADE.downsampleFCS}}
}
