% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_metadata.R
\name{insert_metadata}
\alias{insert_metadata}
\title{Insert metadata to SummarizedExperiment object}
\usage{
insert_metadata(
  SE,
  meta,
  add_column,
  meta_column,
  matching_column,
  fix_metadata = TRUE,
  output_folder = NULL,
  SE_name = NULL
)
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object or path to \code{SummarizedExperiment} object.}

\item{meta}{Path to csv file that contain the data to insert.}

\item{add_column}{Character vector of any length. It represents the column names of the new columns to be inserted.}

\item{meta_column}{Character vector of the same length as \code{add_column} of the names of the columns in \code{meta} that will be inserted.}

\item{matching_column}{Character vector of length 1. It represents an existing matching column in \code{meta} and \code{rowData()}}

\item{fix_metadata}{Logical. If TRUE (default), the column gets added to \code{metadata} slot as well.}

\item{output_folder}{Path of output directory to recieve the subsetted SummarizedExperiment.}

\item{SE_name}{Name of the new subsetted SummarizedExperiment.}
}
\value{
SummarizedExperiment with added data to \code{rowData()}
}
\description{
Insert metadata to SummarizedExperiment object
}
