% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_sweep.R
\name{k_sweep}
\alias{k_sweep}
\title{Phenograh clustering k parameter sweeps}
\usage{
k_sweep(
  SE,
  k,
  export = TRUE,
  output_folder = ".",
  assay_name = "exprs",
  show.plot = FALSE
)
}
\arguments{
\item{SE}{Path to SummarizedExperiment object created by \code{\link{build_SE}}}

\item{k}{Integer vector represent k values to be tested.}

\item{export}{Logical. If TRUE (default), plot is saved in output_folder.}

\item{output_folder}{Path to folder to receive the output.}

\item{assay_name}{Single character vector of the name of the \code{assay()} slot to be used in clustering.}

\item{show.plot}{Logical. If TRUE final plot is  saved and shown. Default (FALSE) will save the plot in working directory without showing it.}
}
\value{
Silhouette plot for the k values selected.
}
\description{
Test multiple values of k on clustering a given data set. Other clustering parameters are defaulted to simulate Phenograph algorithm.
}
\note{
The performance of \code{k_sweep()} is affected significantly by the number of cores and memory available for computation, and also the size of the dataset.
}
\seealso{
For more information and more metrics to assess clustering parameters; see \href{https://bioconductor.org/packages/devel/bioc/vignettes/bluster/inst/doc/diagnostics.html}{Assorted clustering diagnostics}
}
