% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{metadata_tables}
\alias{metadata_tables}
\title{Extract metadata tables from raw fcs files}
\usage{
metadata_tables(input_folder, output_folder, clean_name = FALSE)
}
\arguments{
\item{input_folder}{Path to folder containing raw fcs files.}

\item{output_folder}{Path to folder to receive the tables.}

\item{clean_name}{If TRUE, the returned marker_info table will have the biological marker names cleaned and readable in column "clean_name". If FALSE (default)
the column will be empty and the analyst will be expected to fill this column manually. The name cleaning is done assuming the conventional and the most common marker naming
during the data acquisition. If not sure, leave it to the default value.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{           csv files exported to output_folder.
                   Analyst are required to fill these tables. For marker_info table, column $marker_class should have only one of the three values c("none", "type", "state), and column
                   $channel_type should only have one of the values c("biological', "non_biological"). Failure to do that will cause errors downsteam.
}\if{html}{\out{</div>}}
}
\description{
Extract two metadata tables from fcs files to be completed by the analyst for downstream analysis. First table is the experimental metadata (experiment_info).
Second table is the markers used metadata (marker_info).
}
\note{
clean_name = TRUE will remove the common used characters in naming channels to make the biomarkers more easily identified. But it
could result in some unclear names occasionally. So, the analyst's attention and edits is recommended after table is exported. Generally, clean-names
should not contain any special characters and should not start with a number.
}
