% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_umap.R
\name{pheno_umap}
\alias{pheno_umap}
\title{PhenoGraph clustering and UMAP dimensionality reduction}
\usage{
pheno_umap(
  SE,
  marker_info,
  k = 30,
  n = 15,
  export = FALSE,
  output_folder = ".",
  assay_name = "exprs",
  ...
)
}
\arguments{
\item{SE}{SummarizedExperiment object or path to \code{SummarizedExperiment} object}

\item{marker_info}{Path to marker_info table; one of the csv files generated by \code{\link{metadata_tables}}.}

\item{k}{The number of nearest-neighbours to be used in building PhenoGraph.}

\item{n}{A umap parameter that controls how UMAP balances local versus global structure in the data.}

\item{export}{If \code{TRUE}, \code{SummarizedExperiment} object is exported and saved in the specified destination.}

\item{output_folder}{Destination folder to recieve the outcome. Working directory is the default.}

\item{assay_name}{Single character vector of the name of the \code{assay()} slot to be used in clustering.}

\item{...}{Additional arguments to \code{\link[Rphenograph]{Rphenograph}} and \code{\link[umap]{umap}}.}
}
\value{
SummarizedExperiment object with a cluster_ID column and UMAP coordinates added. Data can be accessed through the \code{rowData()} accessor.
}
\description{
Assigns a phenograph cluster and calculate UMAP axes coordinates for cytometry single-cell data from cellular protein marker expression values,
}
\references{
\href{https://www.nature.com/articles/nbt.4314}{UMAP}
\href{https://www.cell.com/cell/fulltext/S0092-8674(15)00637-6?_returnURL=https\%3A\%2F\%2Flinkinghub.elsevier.com\%2Fretrieve\%2Fpii\%2FS0092867415006376\%3Fshowall\%3Dtrue}{PhenoGraph}
}
\seealso{
\href{https://bioconductor.org/packages/devel/bioc/vignettes/SummarizedExperiment/inst/doc/SummarizedExperiment.html}{Assorted clustering diagnostics}.
}
