% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stainQC.R
\name{plot_stainQC}
\alias{plot_stainQC}
\title{Plot histograms for visual inspection of batch controls for single-cell cytometry data}
\usage{
plot_stainQC(input_folder, pattern, output_folder)
}
\arguments{
\item{input_folder}{Path to folder containing fcs files.}

\item{pattern}{A common string to look for among fcs files in the \code{input_folder} e.g. "control" , "healthy_donor".etc.
It could also be a regular expression.}

\item{output_folder}{Path to folder to save the transformed fcs files.}
}
\value{
Histogram plots of all channels in fcs files in all control files.
}
\description{
Plot histograms of data channels in fcs files for visual inspection and comparison between batch controls.
}
