% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_clus_hm.R
\name{cyto_clust_hm}
\alias{cyto_clust_hm}
\title{Heatmap for clustered single-cell cytometry data}
\usage{
cyto_clust_hm(SE, output_folder = ".", assay_name = "exprs", ...)
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object or path to \code{SummarizedExperiment} object.}

\item{output_folder}{Path to folder to recieve pdf file of the heatmap.}

\item{assay_name}{Name of \code{assay()} slot to be used in generating the heatmap.}

\item{...}{Extra arguments for \code{pheatmap()}}
}
\value{
Heatmap clusters as rows and median protein expression as column.
}
\description{
A wrapper around \code{\link{pheatmap}} package to facilitate  generation of heatmap for clustered single-cell cytoemetry data.
}
\note{
The protein expression is scaled across the clusters.
}
